VLCJ 4 migration

+++ src/main/java/cz/encircled/eplayer/service/MediaService.java
@@ -19,7 +19,7 @@ public interface MediaService {
      */
     void play(@NotNull MediaType p);
 
-    int getMediaLength();
+    long getMediaLength();
 
     long getCurrentTime();
